+++ examples/greet.ts
@@ -1,4 +1,12 @@
-import { createBot, hashtag, Image, link, mention, text } from "@fedify/botkit";
+import {
+  createBot,
+  customEmoji,
+  hashtag,
+  Image,
+  link,
+  mention,
+  text,
+} from "@fedify/botkit";
 import { DenoKvMessageQueue, DenoKvStore } from "@fedify/fedify/x/denokv";
 
 const kv = await Deno.openKv();
@@ -25,9 +33,18 @@ const bot = createBot<void>({
   pages: { color: "green" },
 });
 
+const emojis = bot.addCustomEmojis({
+  botkit: {
+    type: "image/png",
+    file: `${import.meta.dirname}/../docs/public/favicon-192x192.png`,
+  },
+});
+
 bot.onFollow = async (session, followRequest) => {
   await session.publish(
-    text`Thanks for following me, ${followRequest.follower}!`,
+    text`Thanks for following me, ${followRequest.follower}! ${
+      customEmoji(emojis.botkit)
+    }`,
     {
       visibility: "direct",
       attachments: [
@@ -46,15 +63,20 @@ bot.onFollow = async (session, followRequest) => {
 };
 
 bot.onUnfollow = async (session, follower) => {
-  await session.publish(text`Goodbye, ${follower}!`, {
-    visibility: "direct",
-  });
+  await session.publish(
+    text`Goodbye, ${follower}! ${customEmoji(emojis.botkit)}`,
+    { visibility: "direct" },
+  );
 };
 
 bot.onReply = async (session, message) => {
   const botUri = session.actorId.href;
   if (message.mentions.some((a) => a.id?.href === botUri)) return;
-  await message.reply(text`Thanks for your reply, ${message.actor}!`);
+  await message.reply(
+    text`Thanks for your reply, ${message.actor}! ${
+      customEmoji(emojis.botkit)
+    }`,
+  );
 };
 
 bot.onMention = async (_session, message) => {
@@ -64,9 +86,9 @@ bot.onMention = async (_session, message) => {
 const session = bot.getSession(Deno.env.get("ORIGIN") ?? "http://localhost");
 setInterval(async () => {
   const message = await session.publish(
-    text`Hi, folks! It's a minutely greeting. It will be deleted in 30 seconds.  ${
-      hashtag("greet")
-    }`,
+    text`Hi, folks! It's a minutely greeting. It will be deleted in 30 seconds. ${
+      customEmoji(emojis.botkit)
+    } ${hashtag("greet")}`,
   );
   setTimeout(async () => {
     await message.delete();