+++ src/components/Message.tsx
@@ -16,7 +16,13 @@
 /** @jsx react-jsx */
 /** @jsxImportSource @hono/hono/jsx */
 import { LanguageString } from "@fedify/fedify/runtime";
-import { Document, getActorHandle, Image, Link } from "@fedify/fedify/vocab";
+import {
+  Document,
+  Emoji,
+  getActorHandle,
+  Image,
+  Link,
+} from "@fedify/fedify/vocab";
 import type { MessageClass } from "../message.ts";
 import type { Session } from "../session.ts";
 
@@ -41,6 +47,16 @@ export async function Message({ session, message }: MessageProps) {
   });
   const authorHandle = author == null ? null : await getActorHandle(author);
   const attachments = await Array.fromAsync(message.getAttachments());
+  const tags = await Array.fromAsync(message.getTags());
+  const customEmojis: Record<string, string> = {};
+  for (const tag of tags) {
+    if (!(tag instanceof Emoji) || tag.name == null) continue;
+    const icon = await tag.getIcon();
+    if (icon?.url == null) continue;
+    const url = icon.url instanceof Link ? icon.url.href?.href : icon.url.href;
+    if (url == null) continue;
+    customEmojis[tag.name.toString()] = url;
+  }
   return (
     <article>
       <header>
@@ -71,7 +87,9 @@ export async function Message({ session, message }: MessageProps) {
           : <em>(Deleted account)</em>}
       </header>
       <div
-        dangerouslySetInnerHTML={{ __html: `${message.content}` }}
+        dangerouslySetInnerHTML={{
+          __html: renderCustomEmojis(`${message.content}`, customEmojis),
+        }}
         lang={message.content instanceof LanguageString
           ? message.content.language.compact()
           : undefined}
@@ -112,4 +130,30 @@ export async function Message({ session, message }: MessageProps) {
       </footer>
     </article>
   );
+}
+
+const HTML_ELEMENT_REGEXP = /<\/?[^>]+>/g;
+const CUSTOM_EMOJI_REGEXP = /:([a-z0-9_-]+):/gi;
+
+export function renderCustomEmojis(
+  html: string,
+  emojis: Record<string, string>,
+): string {
+  let result = "";
+  let index = 0;
+  for (const match of html.matchAll(HTML_ELEMENT_REGEXP)) {
+    result += replaceEmojis(html.substring(index, match.index));
+    result += match[0];
+    index = match.index + match[0].length;
+  }
+  result += replaceEmojis(html.substring(index));
+  return result;
+
+  function replaceEmojis(html: string): string {
+    return html.replaceAll(CUSTOM_EMOJI_REGEXP, (match) => {
+      const emoji = emojis[match] ?? emojis[match.replace(/^:|:$/g, "")];
+      if (emoji == null) return match;
+      return `<img src="${emoji}" alt="${match}" style="height: 1em">`;
+    });
+  }
 }